+++ statistics-node/statistics-node-all-he/src/main/java/com/dwarfeng/statistics/node/all/he/configuration/ExceptionCodeOffsetConfiguration.java
@@ -1,0 +1,51 @@
+package com.dwarfeng.statistics.node.all.he.configuration;
+
+import com.dwarfeng.statistics.sdk.util.ServiceExceptionCodes;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Configuration;
+
+import javax.annotation.PostConstruct;
+
+@Configuration
+public class ExceptionCodeOffsetConfiguration {
+
+    @Value("${com.dwarfeng.statistics.statistics.exception_code_offset}")
+    private int exceptionCodeOffset;
+    @Value("${com.dwarfeng.statistics.statistics.exception_code_offset.subgrade}")
+    private int subgradeExceptionCodeOffset;
+    @Value("${com.dwarfeng.statistics.statistics.exception_code_offset.spring_telqos}")
+    private int springTelqosExceptionCodeOffset;
+    @Value("${com.dwarfeng.statistics.statistics.exception_code_offset.spring_terminator}")
+    private int springTerminatorExceptionCodeOffset;
+    @Value("${com.dwarfeng.statistics.statistics.exception_code_offset.dwarfeng_datamark}")
+    private int dwarfengDatamarkExceptionCodeOffset;
+    @Value("${com.dwarfeng.statistics.statistics.exception_code_offset.dcti}")
+    private int dctiExceptionCodeOffset;
+    @Value("${com.dwarfeng.statistics.statistics.exception_code_offset.dwarfeng_dct}")
+    private int dwarfengDctExceptionCodeOffset;
+
+    @PostConstruct
+    public void init() {
+        ServiceExceptionCodes.setExceptionCodeOffset(
+                exceptionCodeOffset
+        );
+        com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes.setExceptionCodeOffset(
+                subgradeExceptionCodeOffset
+        );
+        com.dwarfeng.springtelqos.sdk.util.ServiceExceptionCodes.setExceptionCodeOffset(
+                springTelqosExceptionCodeOffset
+        );
+        com.dwarfeng.springterminator.sdk.util.ServiceExceptionCodes.setExceptionCodeOffset(
+                springTerminatorExceptionCodeOffset
+        );
+        com.dwarfeng.datamark.sdk.util.ServiceExceptionCodes.setExceptionCodeOffset(
+                dwarfengDatamarkExceptionCodeOffset
+        );
+        com.dwarfeng.dcti.sdk.util.ServiceExceptionCodes.setExceptionCodeOffset(
+                dctiExceptionCodeOffset
+        );
+        com.dwarfeng.dct.sdk.util.ServiceExceptionCodes.setExceptionCodeOffset(
+                dwarfengDctExceptionCodeOffset
+        );
+    }
+}